+++ AutomatedLabDefinition/AutomatedLabDefinition.psm1
@@ -2034,7 +2034,7 @@ function Add-LabMachineDefinition
 
         if ((-not [string]::IsNullOrWhiteSpace($SshPublicKeyPath) -and [string]::IsNullOrWhiteSpace($SshPrivateKeyPath)) -or ([string]::IsNullOrWhiteSpace($SshPublicKeyPath) -and -not [string]::IsNullOrWhiteSpace($SshPrivateKeyPath)))
         {
-            Write-ScreenInfo -Type Warning -Message "Both SshPublicKeyPath and SshPrivateKeyPath need to be used to successfully remote to Linux VMs"
+            Write-ScreenInfo -Type Warning -Message "Both SshPublicKeyPath and SshPrivateKeyPath need to be used to successfully remote to Linux VMs (Host Windows, Engine Hyper-V) and Windows VMs (Host Linux/WSL, Engine Azure)"
         }
 
         if ($AzureProperties)
@@ -2111,29 +2111,29 @@ function Add-LabMachineDefinition
         $machine.OrganizationalUnit = $OrganizationalUnit
         $script:machines.Add($machine)
 
-        if ($OperatingSystem.OperatingSystemType -eq 'Windows' -and $SshPublicKeyPath)
+        if ($script:lab.DefaultVirtualizationEngine -eq 'HyperV' -and $OperatingSystem.OperatingSystemType -eq 'Windows' -and $SshPublicKeyPath)
         {
-            Write-ScreenInfo -Message "SSH Keys are ignored on Windows for the time being. Why not contribute to AutomatedLab and add the configuration of an ssh server?"
+            Write-ScreenInfo -Message "SSH Keys are ignored on Hyper-V and for Windows targets for the time being. Why not contribute to AutomatedLab and add the configuration of an ssh server?"
         }
-        elseif ($OperatingSystem.OperatingSystemType -eq 'Linux' -and $SshPublicKeyPath -and -not (Test-Path -Path $SshPublicKeyPath))
+        elseif (($script:lab.DefaultVirtualizationEngine -eq 'Azure' -or $OperatingSystem.OperatingSystemType -eq 'Linux') -and $SshPublicKeyPath -and -not (Test-Path -Path $SshPublicKeyPath))
         {
             throw "$SshPublicKeyPath does not exist. Rethink your decision."
         }
-        elseif ($OperatingSystem.OperatingSystemType -eq 'Linux' -and $SshPublicKeyPath)
+        elseif (($script:lab.DefaultVirtualizationEngine -eq 'Azure' -or $OperatingSystem.OperatingSystemType -eq 'Linux') -and $SshPublicKeyPath)
         {
             $machine.SshPublicKeyPath = $SshPublicKeyPath
             $machine.SshPublicKey = Get-Content -Raw -Path $SshPublicKeyPath
         }
 
-        if ($OperatingSystem.OperatingSystemType -eq 'Windows' -and $SshPrivateKeyPath)
+        if ($script:lab.DefaultVirtualizationEngine -eq 'HyperV' -and $OperatingSystem.OperatingSystemType -eq 'Windows' -and $SshPrivateKeyPath)
         {
-            Write-ScreenInfo -Message "SSH Keys are ignored on Windows for the time being. Why not contribute to AutomatedLab and add the configuration of an ssh server?"
+            Write-ScreenInfo -Message "SSH Keys are ignored on Hyper-V and for Windows targets for the time being. Why not contribute to AutomatedLab and add the configuration of an ssh server?"
         }
-        elseif ($OperatingSystem.OperatingSystemType -eq 'Linux' -and $SshPrivateKeyPath -and -not (Test-Path -Path $SshPrivateKeyPath))
+        elseif (($script:lab.DefaultVirtualizationEngine -eq 'Azure' -or $OperatingSystem.OperatingSystemType -eq 'Linux') -and $SshPrivateKeyPath -and -not (Test-Path -Path $SshPrivateKeyPath))
         {
             throw "$SshPrivateKeyPath does not exist. Rethink your decision."
         }
-        elseif ($OperatingSystem.OperatingSystemType -eq 'Linux' -and $SshPrivateKeyPath)
+        elseif (($script:lab.DefaultVirtualizationEngine -eq 'Azure' -or $OperatingSystem.OperatingSystemType -eq 'Linux') -and $SshPrivateKeyPath)
         {
             $machine.SshPrivateKeyPath = $SshPrivateKeyPath
         }